requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
streamlit
openai
python-dotenv
pymupdf

